requirements.txt
flask==2.0.1
groq==0.1.1             
python-dotenv==0.19.2   
requests==2.25.1        
beautifulsoup4==4.9.3   
